+++ app/src/main/java/nextstep/jwp/http/message/response/Response.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.http.message.response;
 
-import nextstep.jwp.http.message.element.cookie.ProxyCookie;
+import nextstep.jwp.http.message.element.cookie.ProxyHttpCookie;
 
 public interface Response {
     String asString();
-    void setCookies(ProxyCookie cookie);
+    void setCookies(ProxyHttpCookie cookie);
 }
