+++ app/src/test/java/nextstep/jwp/http/message/element/CookieTest.java
@@ -1,6 +1,6 @@
 package nextstep.jwp.http.message.element;
 
-import nextstep.jwp.http.message.element.cookie.CookieImpl;
+import nextstep.jwp.http.message.element.cookie.HttpCookie;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -15,7 +15,7 @@ class CookieTest {
     @DisplayName("쿠키가 존재하면 쿠키를 추출한다.")
     @Test
     void get_exist() {
-        CookieImpl cookie = new CookieImpl("test1=1; test2=2");
+        HttpCookie cookie = new HttpCookie("test1=1; test2=2");
 
         assertThat(cookie.get("test1")).get().isEqualTo("1");
         assertThat(cookie.get("test2")).get().isEqualTo("2");
@@ -23,7 +23,7 @@ class CookieTest {
 
     @Test
     void size() {
-        CookieImpl cookie = new CookieImpl("test1=1; test2=2");
+        HttpCookie cookie = new HttpCookie("test1=1; test2=2");
 
         assertThat(cookie.size()).isEqualTo(2);
     }
